DataDome/python/datadome-slider/requirements.txt
capsolver
python-dotenv
